implement locale detection for wayland. (#2059)
