+++ druid-shell/src/backend/shared/linux/mod.rs
@@ -1,0 +1,2 @@
+// environment based utilities
+pub mod env;
